Reverting change to preallocate _result
It seems that putting the result of the CspRecvPort.recv into a pre-allocated numpy _result leads to a hang due to shared memory between processes, specifically in cases where the user code grabs data from a Monitor. Reverting that change for now, though in the future it would be good to find a way to effectively preallocate to avoid memory allocations during every timestep.